feature(fix): separate loading fonts from home component
